infrastructure-fix/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
pytest==7.3.1
boto3==1.26.148
python-dotenv==1.0.0
aws-lambda-powertools==2.18.0
aws-xray-sdk==2.12.0

